python-kafka-and-tooling/basic-example/pyproject.toml
[tool.poetry]
name = "python-kafka-fun"
version = "0.1.0"
description = ""
authors = ["JangasCodingplace <janisgoesser92@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
confluent-kafka = "^2.1.0"

[tool.poetry.dev-dependencies]
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
